backend/requirements.txt
fastapi==0.69.0
pydantic==1.9.0
uvicorn==0.16.0
mysql-connector-python==8.0.27